Move FeatureRegistry to a submodule wagtail.core.rich_text.feature_registry
